requirements.txt
aiodns==1.1.1
aiohttp==3.0.5
alembic==0.8.10
amqp==2.2.2
apache-airflow==1.9.0
asn1crypto==0.24.0
async-timeout==2.0.0
attrs==17.4.0
Babel==2.5.3
bcrypt==3.1.4
billiard==3.5.0.3
bleach==2.1.2
boto==2.48.0
boto3==1.6.3
botocore==1.9.3
cached-property==1.3.1
celery==4.1.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
configparser==3.5.0
croniter==0.3.20
cryptography==2.1.4
dill==0.2.7.1
docker==2.7.0
docker-compose==1.19.0
docker-pycreds==0.2.2
dockerpty==0.4.1
docopt==0.4.0
docutils==0.14
filechunkio==1.8
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
flower==0.9.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.8
google-api-python-client==1.6.5
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
idna-ssl==1.0.0
itsdangerous==0.24
itunes-iap==2.5.2
Jinja2==2.8.1
jmespath==0.9.3
jsonschema==2.6.0
kombu==4.1.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
mandrill==1.0.57
Markdown==2.6.11
MarkupSafe==1.0
multidict==4.1.0
numpy==1.14.1
oauth2client==4.1.2
ordereddict==1.1
pandas==0.22.0
prettyexc==0.6.0
psutil==4.4.2
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycares==2.3.0
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
setproctitle==1.1.10
six==1.11.0
slackclient==1.1.2
smmap2==2.0.3
SQLAlchemy==1.2.4
tabulate==0.7.7
texttable==0.9.1
thrift==0.11.0
tornado==4.5.3
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
websocket-client==0.47.0
Werkzeug==0.14.1
WTForms==2.1
yarl==1.1.1
zope.deprecation==4.3.0
